new changes

+++ E-TicketMovies/Areas/Admin/Controllers/ActorController.cs
@@ -1,12 +1,14 @@
 ﻿using E_TicketMovies.Models;
 using E_TicketMovies.Repositories;
 using E_TicketMovies.Repositories.IRepositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace E_TicketMovies.Areas.Admin.Controllers
 {
     [Area("Admin")]
+    [Authorize(Roles = "Admin,SuperAdmin")]
     public class ActorController : Controller
     {
         IActorRepository actorRepository;
@@ -18,10 +20,25 @@ namespace E_TicketMovies.Areas.Admin.Controllers
             this.movieRepository = movieRepository;
             this.actorMovieRepository = actorMovieRepository;
         }
-        public IActionResult Index()
+        public IActionResult Index(string? query, int page =1)
         {
-            var actors = actorRepository.Get().ToList();
-            return View(actors);
+            var actors = actorRepository.Get();
+            if (query!=null)
+            {
+                actors = actorRepository.Get(e=>e.FirstName.Contains(query) || e.LastName.Contains(query) || e.Bio.Contains(query));
+            }
+            int totalCount = actors.Count();
+            int pageSize = 8;
+            int totalPages = (int)Math.Ceiling((double)totalCount / pageSize);
+
+            if (page > totalPages && totalPages > 0)
+                return RedirectToAction("NotFoundPage", "Home", new { area = "End User" });
+
+            actors = actors.Skip((page - 1) * pageSize).Take(pageSize);
+
+            ViewBag.totalPages = totalPages;
+
+            return View(actors.ToList());
         }
         [HttpGet]
         public IActionResult Create()
@@ -29,16 +46,16 @@ namespace E_TicketMovies.Areas.Admin.Controllers
             var movies = movieRepository.Get().ToList();
             ViewBag.Movies = movies;
 
-            return View();
+            return View(new Actor ());
         }
         [HttpPost]
         public IActionResult Create(Actor actor, IFormFile ProfilePicture, List<int> moviesId)
         {
-  
+            var movies = movieRepository.Get().ToList();
+            ViewBag.Movies = movies;
+
             if (ModelState.IsValid)
-            {
-                if (ProfilePicture != null && ProfilePicture.Length > 0)
-                {
+            { 
                     var imageName = Guid.NewGuid().ToString() + Path.GetExtension(ProfilePicture.FileName);
                     var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cast", imageName);
                     using (var stream = System.IO.File.Create(filePath))
@@ -46,7 +63,7 @@ namespace E_TicketMovies.Areas.Admin.Controllers
                         ProfilePicture.CopyTo(stream);
                     }
                     actor.ProfilePicture = imageName;
-                }
+                
             
                 actorRepository.Create(actor);
                 actorRepository.Commit();
@@ -71,7 +88,8 @@ namespace E_TicketMovies.Areas.Admin.Controllers
 
                 return RedirectToAction(nameof(Index));
             }
-            return RedirectToAction("NotFoundPage", "Home");
+            
+            return View (actor);
 
         }
         [HttpGet]
@@ -82,19 +100,18 @@ namespace E_TicketMovies.Areas.Admin.Controllers
             ViewBag.Movies = movies;
             var selectedMovies = actorMovieRepository.Get(e=>e.ActorId == id).Select(e=>e.MovieId).ToList();
             ViewBag.SelectedMovies = selectedMovies;
-            return View(actor);
+            return View(new Actor());
         }
         [HttpPost]
         public IActionResult Edit(Actor actor, IFormFile ProfilePicture, List<int> moviesId)
         {
-            var updatedActor = actorRepository.GetOne(e => e.Id == actor.Id ,tracked: false);
+            var movies = movieRepository.Get().ToList();
+            ViewBag.Movies = movies;
+            var updatedActor = actorRepository.GetOne(e => e.Id == actor.Id, tracked: false);
             var actorMovies = actorMovieRepository.Get(e => e.ActorId == actor.Id).ToList();
 
-            if (updatedActor != null)
-            {
                 if (ModelState.IsValid)
                 {
-                   
                     if (ProfilePicture != null)
                     {
                         var imageName = Guid.NewGuid().ToString() + Path.GetExtension(ProfilePicture.FileName);
@@ -105,12 +122,11 @@ namespace E_TicketMovies.Areas.Admin.Controllers
                             ProfilePicture.CopyTo(stream);
                         }
 
-                        
-                        var oldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cast", updatedActor.ProfilePicture);
-                        if (System.IO.File.Exists(oldPath))
-                        {
-                            System.IO.File.Delete(oldPath);
-                        }
+                            var oldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cast", updatedActor.ProfilePicture);
+                            if (System.IO.File.Exists(oldPath))
+                            {
+                                System.IO.File.Delete(oldPath);
+                            }
 
                         actor.ProfilePicture = imageName;
                     }
@@ -124,7 +140,7 @@ namespace E_TicketMovies.Areas.Admin.Controllers
 
                     // handlle selected movies
                     var existingActorMovies = actorMovieRepository.Get(e => e.ActorId == actor.Id).ToList();
-                    var existingMovieIds = existingActorMovies.Select(e=>e.MovieId).ToList();
+                    var existingMovieIds = existingActorMovies.Select(e => e.MovieId).ToList();
                     // remove unselected movies
                     var moviesToRemove = existingActorMovies.Where(e => !moviesId.Contains(e.MovieId)).ToList();
                     foreach (var movie in moviesToRemove)
@@ -146,12 +162,9 @@ namespace E_TicketMovies.Areas.Admin.Controllers
                     }
                     return RedirectToAction(nameof(Index));
                 }
-                else
-                {
-                    actor.ProfilePicture = updatedActor.ProfilePicture;
-                }
-            }
-            return RedirectToAction("NotFoundPage", "Home");
+               
+            
+            return View(actor);
         }
 
 
@@ -171,12 +184,14 @@ namespace E_TicketMovies.Areas.Admin.Controllers
             }
             if (deletedActor != null)
             {
-                var oldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cast", deletedActor.ProfilePicture);
-                if (System.IO.File.Exists(oldPath))
+                if (deletedActor.ProfilePicture != null)
                 {
-                    System.IO.File.Delete(oldPath);
+                    var oldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cast", deletedActor.ProfilePicture);
+                    if (System.IO.File.Exists(oldPath))
+                    {
+                        System.IO.File.Delete(oldPath);
+                    }
                 }
-
                 actorRepository.Delete(deletedActor);
                 actorRepository.Commit();
                 return RedirectToAction(nameof(Index));
