+++ E-TicketMovies/Areas/Admin/Controllers/CinemaController.cs
@@ -1,11 +1,13 @@
 ﻿using E_TicketMovies.Models;
 using E_TicketMovies.Repositories.IRepositories;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 
 namespace E_TicketMovies.Areas.Admin.Controllers
 {
     [Area("Admin")]
+    [Authorize(Roles = "Admin,SuperAdmin")]
 
     public class CinemaController : Controller
     {
@@ -15,10 +17,26 @@ namespace E_TicketMovies.Areas.Admin.Controllers
             this.cinemaRepository = cinemaRepository;
             
         }
-        public IActionResult Index()
+        public IActionResult Index(string? query, int page)
         {
-            var cinema = cinemaRepository.Get().ToList();
-            return View(cinema);
+            var cinema = cinemaRepository.Get();
+
+            if (query != null)
+            {
+                cinema = cinemaRepository.Get(e=>e.Name.Contains(query));
+            }
+            int totalCount = cinema.Count();
+            int pageSize = 3;
+            int totalPages = (int)Math.Ceiling((double)totalCount / pageSize);
+
+            if (page > totalPages && totalPages > 0)
+                return RedirectToAction("NotFoundPage", "Home", new { area = "End User" });
+
+            cinema = cinema.Skip((page - 1) * pageSize).Take(pageSize);
+
+            ViewBag.totalPages = totalPages;
+
+            return View(cinema.ToList());
         }
         [HttpGet]
         public IActionResult Create()
@@ -54,34 +72,38 @@ namespace E_TicketMovies.Areas.Admin.Controllers
             return View(cinema);
         }
         [HttpPost]
-        public IActionResult Edit(Cinema cinema, IFormFile? CinemaLogo)
+        public IActionResult Edit(Cinema cinema, IFormFile CinemaLogo)
         {
-            var updatedCinema = cinemaRepository.GetOne(e => e.Id == cinema.Id,tracked:false);
-            if (updatedCinema != null) { 
+            var updatedCinema = cinemaRepository.GetOne(e => e.Id == cinema.Id, tracked: false);
 
             if (ModelState.IsValid)
             {
-                var imageName = Guid.NewGuid().ToString() + Path.GetExtension(CinemaLogo?.FileName);
-                var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cinema", imageName);
-                using (var stream = System.IO.File.Create(filePath))
+                if (CinemaLogo != null)
                 {
-                    CinemaLogo?.CopyTo(stream);
+                    var imageName = Guid.NewGuid().ToString() + Path.GetExtension(CinemaLogo?.FileName);
+                    var filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cinema", imageName);
+                    using (var stream = System.IO.File.Create(filePath))
+                    {
+                        CinemaLogo.CopyTo(stream);
+                    }
+                    
+                    var oldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cinema", updatedCinema.CinemaLogo);
+                    if (System.IO.File.Exists(oldPath))
+                    {
+                        System.IO.File.Delete(oldPath);
+                    }
+                    cinema.CinemaLogo = imageName;
                 }
-                cinema.CinemaLogo = imageName;
-
-                var oldPath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot\\images\\cinema", updatedCinema.CinemaLogo);
-                if (System.IO.File.Exists(oldPath))
+                else
                 {
-                    System.IO.File.Delete(oldPath);
+                    cinema.CinemaLogo = updatedCinema.CinemaLogo;
                 }
+            
                 cinemaRepository.Update(cinema);
                 cinemaRepository.Commit();
-            }
-            else
-            {
-                cinema.CinemaLogo = updatedCinema.CinemaLogo;
-            }
+
                 return RedirectToAction(nameof(Index));
+   
             }
             return View(cinema);
         }
